+++ src/gui/app.rs
@@ -1956,6 +1956,15 @@ impl Application for App {
                 search.change.choice = filter;
                 Command::none()
             }
+            Message::EditedSearchFilterManifest(filter) => {
+                let search = if self.screen == Screen::Backup {
+                    &mut self.backup_screen.log.search
+                } else {
+                    &mut self.restore_screen.log.search
+                };
+                search.manifest.choice = filter;
+                Command::none()
+            }
             Message::EditedSortKey { screen, value } => {
                 match screen {
                     Screen::Backup => {