+++ src/gui/search.rs
@@ -8,6 +8,7 @@ use crate::{
         widget::{checkbox, pick_list, text, Column, Element, IcedParentExt, Row},
     },
     lang::TRANSLATOR,
+    resource::manifest::Manifest,
     scan::{
         game_filter::{self, FilterKind},
         Duplication, ScanInfo,
@@ -28,6 +29,7 @@ pub struct FilterComponent {
     pub completeness: Filter<game_filter::Completeness>,
     pub enablement: Filter<game_filter::Enablement>,
     pub change: Filter<game_filter::Change>,
+    pub manifest: Filter<game_filter::Manifest>,
 }
 
 fn template<'a, T: 'static + Default + Copy + Eq + PartialEq + ToString>(
@@ -50,10 +52,28 @@ fn template<'a, T: 'static + Default + Copy + Eq + PartialEq + ToString>(
         .into()
 }
 
+fn template_with_label<T: 'static + Default + Clone + Eq + PartialEq + ToString>(
+    filter: &Filter<T>,
+    label: String,
+    kind: FilterKind,
+    options: Vec<T>,
+    message: fn(T) -> Message,
+) -> Element {
+    Row::new()
+        .spacing(10)
+        .align_items(Alignment::Center)
+        .push(checkbox(label, filter.active, move |enabled| {
+            Message::ToggledSearchFilter { filter: kind, enabled }
+        }))
+        .push(pick_list(options, Some(filter.choice.clone()), message))
+        .into()
+}
+
 impl FilterComponent {
     pub fn qualifies(
         &self,
         scan: &ScanInfo,
+        manifest: &Manifest,
         enabled: bool,
         duplicated: Duplication,
         show_deselected_games: bool,
@@ -66,8 +86,14 @@ impl FilterComponent {
         let complete = !self.completeness.active || self.completeness.choice.qualifies(scan);
         let enable = !show_deselected_games || !self.enablement.active || self.enablement.choice.qualifies(enabled);
         let changed = !self.change.active || self.change.choice.qualifies(scan);
+        let manifest = !self.manifest.active
+            || manifest
+                .0
+                .get(&scan.game_name)
+                .map(|game| self.manifest.choice.qualifies(game))
+                .unwrap_or_default();
 
-        fuzzy && unique && complete && changed && enable
+        fuzzy && unique && complete && changed && enable && manifest
     }
 
     pub fn toggle_filter(&mut self, filter: FilterKind, enabled: bool) {
@@ -76,10 +102,17 @@ impl FilterComponent {
             FilterKind::Completeness => self.completeness.active = enabled,
             FilterKind::Enablement => self.enablement.active = enabled,
             FilterKind::Change => self.change.active = enabled,
+            FilterKind::Manifest => self.manifest.active = enabled,
         }
     }
 
-    pub fn view(&self, screen: Screen, histories: &TextHistories, show_deselected_games: bool) -> Option<Element> {
+    pub fn view(
+        &self,
+        screen: Screen,
+        histories: &TextHistories,
+        show_deselected_games: bool,
+        manifests: Vec<game_filter::Manifest>,
+    ) -> Option<Element> {
         if !self.show {
             return None;
         }
@@ -128,6 +161,19 @@ impl FilterComponent {
                             )
                         }),
                 )
+                .push_if(manifests.len() > 1, || {
+                    Row::new()
+                        .padding([0, 20, 20, 20])
+                        .spacing(20)
+                        .align_items(Alignment::Center)
+                        .push(template_with_label(
+                            &self.manifest,
+                            TRANSLATOR.source_field(),
+                            FilterKind::Manifest,
+                            manifests,
+                            Message::EditedSearchFilterManifest,
+                        ))
+                })
                 .into(),
         )
     }